Replace `adm-zip` with `yauzl`
`adm-zip` was a poor choice, it doesn't handle errors at all leading to
obscure error messages that I had to go digging into the source code to
figure out what was going on.

`yauzl` on the other hand is very strict aginst the spec and is the most
popular / battle-tested unzip module on npm.