Inline `promisify()` polyfill
These runtimes should not depend on external dependencies, so
`util.promisify` module has been removed.

This is due to compatability with `@zeit/ncc`, where because this
`bootstrap.js` file is not actually imported by the now-cli code, the
external dependency is not available in the resulting bundle.

It coincidentally worked during `now dev` development from within the
repo, because the `util.promisify` module was present upwards of the
`dist` dir, but the issue surfaced once you zipped up the `dist` dir and
extracted it elsewhere.