WIP `pkg` support
The main change is that everything from the `runtime` dir is copied into
the cache dir, so that the `bootstrap` file actually exists and can be
executed by `execve()`.

The `pkg` test is not yet passing because the `copyfile` syscall is not
yet implemented in `pkg`, so the fs-extra's `copy()` operation fails.

Next up will be to work around that issue using `readFile()` and
`writeFile()` which should work with `pkg` already.