Refactor the runtime cache invalidation strategy to be more bullet-proof
Previously, the cache invalidation strategy was a manual process:
if one of the files in a runtime implementation changed, then the
`version` property was supposed to be updated in the runtime file and
then `fun` would re-initialize the runtime into the cache directory on
the next pass.

This was error-prone because if you forgot to update the `version` property
then the user will end up seeing mismatched behavior, and likely a nasty
error.

The new approach generates a sha256 sum of all the files within the
runtime directory, and if it doesn't match the shasum in the
`$CACHE_DIR/.cache-sha` file then the cached version is stale, and `fun`
will re-initialize the runtime from its sources. This auto-generated
shasum approach takes the human error element out of the equation, and
should be much "safer" going forward.