Wait for API responses to flush before resolving promises
This is an attempt to make sure the invoke response HTTP request by the
runtime is processed before the process gets frozen. In the Go runtime,
the request was not being handled before being frozen in some cases,
which leads to the HTTP request timing out and the process exiting.