Export the `initializeRuntime()` function and allow string runtime names
This will be used by `now dev` to launch the builders in a forked child
process using the same version of `node` that is used by builders in
production, but not actually be a `fun` lambda function.

Also adds the `${cacheDir}/bin` of the runtime to the `PATH` env var in
the fun function so that it has access to the binaries provided there by
default.