Use the `tree-kill` module to shut down lambda functions
In the past, it has been witnessed that `bootstrap` child
processes remain alive even after `fn.destroy()` is invoked.
The circumstances for when/why this happens is uncertain at
this time, however we are using `tree-kill` in Now CLI as well
to handle killing sub-tree child processes that builders create.